refactor(server): move the #106 gate tests to their own file under the 25-it cap
The hygiene gate caps files at 25 its; world-evaluator.test.ts was already
at the cap. The meaning-made boundary tests move to
meaning-made-gate.test.ts (5 its); test inventory regenerated.

audit-tests: PASS — 116 files, 1110 its, no violations.
